fix: Move last functions from StatusConfiguration to Status
This fixes #1535: 'Some "refactoring" broke reading of status settings'
