+++ tests/Status.test.ts
@@ -9,18 +9,23 @@ jest.mock('obsidian');
 window.moment = moment;
 
 describe('Status', () => {
-    describe('default configurations', () => {
-        expect(Status.DONE.configuration.previewText()).toEqual("- [x] Done, next status is ' '. ");
-        expect(Status.EMPTY.configuration.previewText()).toEqual("- [] EMPTY, next status is ''. ");
-        expect(Status.TODO.configuration.previewText()).toEqual("- [ ] Todo, next status is 'x'. ");
+    it('preview text', () => {
+        const configuration = new Status(new StatusConfiguration('P', 'Pro', 'Con', true));
+        expect(configuration.previewText()).toEqual("- [P] Pro, next status is 'Con'. ");
+    });
+
+    it('default configurations', () => {
+        expect(Status.DONE.previewText()).toEqual("- [x] Done, next status is ' '. ");
+        expect(Status.EMPTY.previewText()).toEqual("- [] EMPTY, next status is ''. ");
+        expect(Status.TODO.previewText()).toEqual("- [ ] Todo, next status is 'x'. ");
     });
 
     it('factory methods for default statuses', () => {
-        expect(Status.makeDone().configuration.previewText()).toEqual("- [x] Done, next status is ' '. ");
-        expect(Status.makeEmpty().configuration.previewText()).toEqual("- [] EMPTY, next status is ''. ");
-        expect(Status.makeTodo().configuration.previewText()).toEqual("- [ ] Todo, next status is 'x'. ");
-        expect(Status.makeCancelled().configuration.previewText()).toEqual("- [-] Cancelled, next status is ' '. ");
-        expect(Status.makeInProgress().configuration.previewText()).toEqual("- [/] In Progress, next status is 'x'. ");
+        expect(Status.makeDone().previewText()).toEqual("- [x] Done, next status is ' '. ");
+        expect(Status.makeEmpty().previewText()).toEqual("- [] EMPTY, next status is ''. ");
+        expect(Status.makeTodo().previewText()).toEqual("- [ ] Todo, next status is 'x'. ");
+        expect(Status.makeCancelled().previewText()).toEqual("- [-] Cancelled, next status is ' '. ");
+        expect(Status.makeInProgress().previewText()).toEqual("- [/] In Progress, next status is 'x'. ");
     });
 
     it('should initialize with valid properties', () => {
